@@ -2,12 +2,14 @@
     "name": "homebridge-magic-blue-bulb",
     "version": "1.0.3",
     "description": "Homebridge plugin for the MagicBlue LED Lightbulbs",
-    "main": "index.js",
+    "main": "dist/index.js",
     "files": [
-        "index.js",
-        "rgbConversion.js"
+        "dist"
     ],
     "scripts": {
+        "build": "tsc",
+        "watch": "tsc --watch",
+        "prepublishOnly": "npm run build",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "repository": {
@@ -24,13 +26,17 @@
     },
     "homepage": "https://github.com/lucavb/homebridge-magic-blue-bulb#readme",
     "engines": {
-        "node": ">=0.12.0",
-        "homebridge": ">=0.2.0"
+        "node": ">=14.0.0",
+        "homebridge": ">=1.3.0"
     },
     "dependencies": {
-        "noble": "1.8.0"
+        "noble": "1.9.1"
     },
     "devDependencies": {
-        "prettier": "2.2.1"
+        "prettier": "2.2.1",
+        "typescript": "^4.9.5",
+        "@types/node": "^18.0.0",
+        "@types/noble": "0.0.44",
+        "homebridge": "^1.6.0"
     }
 }
