@@ -5,29 +5,37 @@ http://axonflux.com/handy-rgb-to-hsl-and-rgb-to-hsv-color-model-c
 
 */
 
+export interface RgbColor {
+    r: number;
+    g: number;
+    b: number;
+}
+
 /**
- * Converts an HSV color value to RGB. Conversion formula
+ * Converts an HSL color value to RGB. Conversion formula
  * adapted from http://stackoverflow.com/a/17243070/2061684
  * Assumes h in [0..360], and s and l in [0..100] and
  * returns r, g, and b in [0..255].
  *
- * @param   {Number}  h       The hue
- * @param   {Number}  s       The saturation
- * @param   {Number}  l       The lightness
- * @return  {Array}           The RGB representation
+ * @param   {number}  h       The hue
+ * @param   {number}  s       The saturation
+ * @param   {number}  l       The lightness
+ * @return  {RgbColor}        The RGB representation
  */
-module.exports.hslToRgb = function (h, s, v) {
-    var r, g, b, i, f, p, q, t;
+export function hslToRgb(h: number, s: number, v: number): RgbColor {
+    let r = 0,
+        g = 0,
+        b = 0;
 
     h /= 360;
     s /= 100;
     v /= 100;
 
-    i = Math.floor(h * 6);
-    f = h * 6 - i;
-    p = v * (1 - s);
-    q = v * (1 - f * s);
-    t = v * (1 - (1 - f) * s);
+    const i = Math.floor(h * 6);
+    const f = h * 6 - i;
+    const p = v * (1 - s);
+    const q = v * (1 - f * s);
+    const t = v * (1 - (1 - f) * s);
     switch (i % 6) {
         case 0:
             r = v;
@@ -60,9 +68,13 @@ module.exports.hslToRgb = function (h, s, v) {
             b = q;
             break;
     }
-    var rgb = { r: Math.round(r * 255), g: Math.round(g * 255), b: Math.round(b * 255) };
+    const rgb: RgbColor = {
+        r: Math.round(r * 255),
+        g: Math.round(g * 255),
+        b: Math.round(b * 255),
+    };
     return rgb;
-};
+}
 
 /**
  * Converts an RGB color value to HSL. Conversion formula
@@ -70,25 +82,25 @@ module.exports.hslToRgb = function (h, s, v) {
  * Assumes r, g, and b are in [0..255] and
  * returns h in [0..360], and s and l in [0..100].
  *
- * @param   {Number}  r       The red color value
- * @param   {Number}  g       The green color value
- * @param   {Number}  b       The blue color value
- * @return  {Array}           The HSL representation
+ * @param   {number}  r       The red color value
+ * @param   {number}  g       The green color value
+ * @param   {number}  b       The blue color value
+ * @return  {number[]}        The HSL representation
  */
-module.exports.rgbToHsl = function (r, g, b) {
+export function rgbToHsl(r: number, g: number, b: number): [number, number, number] {
     r /= 255;
     g /= 255;
     b /= 255;
-    var max = Math.max(r, g, b),
-        min = Math.min(r, g, b);
-    var h,
-        s,
-        l = (max + min) / 2;
+    const max = Math.max(r, g, b);
+    const min = Math.min(r, g, b);
+    let h: number;
+    let s: number;
+    const l = (max + min) / 2;
 
-    if (max == min) {
+    if (max === min) {
         h = s = 0; // achromatic
     } else {
-        var d = max - min;
+        const d = max - min;
         s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
         switch (max) {
             case r:
@@ -100,12 +112,14 @@ module.exports.rgbToHsl = function (r, g, b) {
             case b:
                 h = (r - g) / d + 4;
                 break;
+            default:
+                h = 0;
         }
         h /= 6;
     }
 
     h *= 360; // return degrees [0..360]
     s *= 100; // return percent [0..100]
-    l *= 100; // return percent [0..100]
-    return [parseInt(h), parseInt(s), parseInt(l)];
-};
+    const lightness = l * 100; // return percent [0..100]
+    return [parseInt(h.toString()), parseInt(s.toString()), parseInt(lightness.toString())];
+}
